cmd/cueckoo: better repository_dispatch message (#13)
When a repository_dispatch is triggered, the GitHub Actions UI shows
that message as the title in the workflow list. Previously that message
simply read "Normal unity run" for every cueckoo runtrybot --versions
invocation, thereby making them indistinguishable in the list.

Fix that by adding the list of versions supplied as part of the title.

Also bump Go version in the CI matrix.