cmd/cueckoo: make unity configuration optional for runtrybot cmd (#19)
Only the main CUE repository actually uses and runs unity. Other
repositories which might want to leverage the cueckoo runtrybot setup
for their GerritHub setup will not know or care about unity, only the
concept of trybots (i.e. tests in the GitHub mirror).

Therefore, we make the unity configuration optional, and only attempt to
trigger a unity build it if exists.

This could be made more pretty/generic in the future, but for now it
works.