cmd/cueckoo: name trybot trigger trybot
The runtrybot command for cmd/cueckoo does two things: triggers trybots,
and, if the repository is configured against unity, triggers unity,
according to the arguments given.

However, the "type" payload sent when triggering the trybots is
currently "runtrybot". This is rather odd, when the workflow on the
other end is almost certainly called "trybot". Therefore, make this
consistent and name the type "trybot". The cmd/cueckoo command is still
called runtrybot, because that's more descriptive of what the user is
actually doing (even though the term trybot then encompasses both the
trybot workflow and potentially unity).

Also re-add newlines to the end of golden files, because hand editing
leaves the files in that state in any case.

Also remove a stray reference to the mirror command that got deleted
earlier.