cmd/cueckoo: support runtrybot HEAD on backport CLs
When we use Gerrit to cherry-pick a change into a release branch,
both the old and new commits end up sharing the same Change-Id.
For example, see https://review.gerrithub.io/c/cue-lang/cue/+/549247
which was backported as https://review.gerrithub.io/c/cue-lang/cue/+/549325.

When trying to use "cueckoo runtrybot HEAD" on the backporting commit,
I got a generic failure that Gerrit couldn't find that change.

It turns out that this duplicate Change-Id is the reason,
per https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#change-id:

	Gerrit also supports the following identifiers:
	* a Change-Id if it uniquely identifies one change ("I8473b95934b5732ac55d26311a706c9c2bde9940")

In this scenario, our Change-Id does not uniquely identify one change,
so we can't use it in a "find one change" query by ID.
Instead, query for that Change-Id and filter by "open" status,
and expect the result to be exactly one change.

For our two CLs above, the equivalent Change-Id query is as follows,
minus the "is:open" bit to show the two merged results:

	https://review.gerrithub.io/q/change:Ia86e9efecde0225caf0d700bf797a3b2e6cfb2f1

This should result in no changes to existing scenarios where a Change-Id
does uniquely identify a change. One potential change is that we will
now refuse to run the trybots on merged or abandoned CLs, but I reckon
that is a reasonable thing to do in any case.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>