cmd/cueckoo: have importpr automatically add "Closes #N" (#31)
It's very common and frustrating to forget to add "Closes #N" to each PR
that we import. Fix that by making importpr smarter.

Also, set the upstream of the branch within which we create we create
the import commit earlier. That way, if any later commands fail (which
they shouldn't really, especially if we are not rebasing onto the tip of
the target branch) we are left in a good state with respect to the
upstream.