cmd/cueckoo: add short commit hash to "Closes" for importpr
When we import a PR from GitHub we immediately lose track of the
original commit sha because of necessary commit message rewriting.
Currently this means we have no easy way of anyone tracking exactly what
was imported.

Per a suggestion from @mvdan, we can simply add this to the "Closes"
message for now, and later consider writing a message back to the PR
from which we are importing with more information.

For #33.