cmd/cueckoo: use unambiguous change IDs to fetch CLs when possible
In some cases it's not possible, such as when one runs

	cueckoo runtrybot --change 1234

or when one runs

	cueckoo runtrybot HEAD

and HEAD is a detached commit or a branch without an upstream.

However, the most common scenario is HEAD with a branch which is
tracking an upstream branch like master or release-branch.v0.5,
so in those cases make the ID unique.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>