cmd/cueckoo: don't send fully unique change IDs to trybot_dispatch
In an earlier commit, we changed our changeID field so that now it will
be a fully unique change ID with tildes whenever possible.

However, trybot_dispatch expects just the Change-Id as it is in the git
commit message trailer, without any tildes.
Adding tildes will break the workflow per the added docs.

cueckoo only needs the unique ID to retrieve each CL via the Gerrit API.
When sending the payload to trybot_dispatch, continue with the same
non-unique change ID form we had before.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>