cmd/cueckoo: fix a regression in runtrybot --change
A recent commit started using the CurrentRevision field returned by the
Gerrit REST API, which saved us a considerable amount of code.

However, we didn't actually use it to index into the Revisions map,
so we would always try to use rev.revision even when it was unset.

Fix that, make the error message handle empty strings better,
and add docs as I had to remind myself that the revision and commit
values are always commit hashes when not empty.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>