cmd/cueckoo: teach runtrybot to accept abbreviated commits
runtrybot without the --change flag can accept arguments:

* the special value of HEAD
* a git commitish value

HEAD is treated specially by cmd/cueckoo, just as it is git-codereview,
in meaning all the commits from the merge-point of the current branch
with respect to the upstream, to the tip of the current branch.

Otherwise, the commitish arguments should be resolved to a commit.

The go-git package we are using does not appear to easily understand the
abbreviated form of commits.

So to teach cmd/cueckoo runtrybot in particular about abbreviated
commits, first use git rev-parse to expand the commitish argument to the
long hash of a commit.