cmd/cueckoo: fix runtrybot taking a list of commits
In a4e0296 we moved away from using go-git to instead use git itself to
resolve commits etc.

This was tested using the special HEAD argument to runtrybot, but not a
specific list of commits.

a4e0296 broke calling runtrybot with a specific commit.

Fix that.

Tested locally with both HEAD and a list of commits.