cmd/cueckoo: stop triggering the old unity in runtrybot
Now that the cue-unity/unity-private setup includes all the projects
previously tested via cue-unity/unity, the duplicate CI setup
is no longer necessary.

Since we'll land the codereview.cfg change to cue-lang/cue separately,
for the time being, cueckoo continues to load both config entries.
It prefers the "new" entry, which works with today's setup:

    cue-unity: https://github.com/cue-unity/unity
    cue-unity-new: https://github.com/cue-unity/unity-private

Then, when the cue-lang/cue repo is switched to one entry again,
it will fall back to that entry with the new repository URL:

    cue-unity: https://github.com/cue-unity/unity-private

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I418aaa420757cc6c9f668e24ca72ec912bb90e0e