drop cuelang.org/go as a dependency
Its only use in cmd/cueckoo was for errors, which is unnecessary,
given that we don't actually create any custom CUE errors
nor use anything that creates them such as the evaluator.

The only other use was tools.go, for updateTxtarTests,
but that was removed in the upstream cue-lang/cue repo in July 2022.
We will sync the CI setup sooner or later, but for now,
removing the custom command which is not even used is very easy.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ide6777d86228a573887327fbce9511b1d2d39e48