cmd/cueckoo: swap x/build/gerrit with go-gerrit
This brings us in line with our other Gerrit clients in Go:

* services/svc/webhooks/gerritstatusupdater/gerritstatusupdater.go
* unity-private/cmd/unity/cmd/change_resolver.go

Beyond the consistency and ease of maintenance for us,
x/build is tailored for the Go team's needs with build.golang.org,
whereas go-gerrit is a generic client library for everyone else.

The next commit will start using an authenticated Gerrit client,
so it was important to do this first to not write even more code
against x/build when we wanted to switch away from it.

Verified that runtrybot still works via https://cuelang.org/cl/553879.

Fixes #22.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ib8017b642307b028b7873cd127ff083d5260b8d6
Reviewed-on: https://review.gerrithub.io/c/cue-sh/tools/+/557233
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Paul Jolly <paul@myitcv.io>