cmd/cueckoo: start using `git credential` for github auth
We want this for Gerrit, but for consistency and to test it out first,
do it with GitHub first.

Note that we still support the env vars as a fallback for now,
in case anyone struggles with the git credential helper setup.
However, we would recommend to avoid the env vars after this change,
as many git credential helpers like libsecret are safer than env vars.

Once again, verified that runtrybot still works with a dummy CL.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Idc188d2168c168c36afd080bed6349e2d042b463
Reviewed-on: https://review.gerrithub.io/c/cue-sh/tools/+/557234
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>