cmd/cueckoo: do not over-run trybots
In case we already have a trybot result for the requested change, there
is no real point requesting them again, unless in exceptional
circumstances.

Similarly, if we flip to drive the running of trybots from a label
(whether using webhooks or not) we only want to trigger the trybots if
someone hasn't already requested them.

We can use labels in both instances to conditionally trigger the
trybots.

This short-circuit can be overridden by passing the --force (or -f for
short) flag to runtrybot.

Otherwise, the trybots are only triggered if neither the Run-TryBot ==
+1 nor TryBot-Result == +1. Note this creates something of an edge case
if someone is triggering the trybots against a specific commit, because
as far as I can tell a label is connected to a change (implicitly the
latest revision) and not a specific commit (potentially not the latest
revision).

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ia8ad5b61b6e3754918b73cd0aad39b5818c4fab6