cmd/cueckoo: do not over-run trybots
In case we already have a trybot result for the requested change, there
is no real point requesting them again, unless in exceptional
circumstances or we want to force the run.

We can use labels to conditionally trigger the trybots.

This short-circuit can be overridden by passing the --force (or -f for
short) flag to runtrybot.

Otherwise, the trybots are skipped if the revision against which we have
requested the trybots is the current revision, and Run-TryBot == +1.
Otherwise they are run.

Tested against this CL!

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ia8ad5b61b6e3754918b73cd0aad39b5818c4fab6
Reviewed-on: https://review.gerrithub.io/c/cue-sh/tools/+/1169892
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>