cmd/cueckoo: drop --change flag
When being given arguments, use a regular expression to see
if an argument looks like a CL number or full Change-Id string.

It is possible for one to name a git branch to match those forms,
but I don't think that is a likely scenario.
Moreover, the most common git ref use case is either a git commit hash
or HEAD to test all commits from the branchpoint.

This bites me every time I want to run the trybots on someone else's CL,
since I give their CL number or Change-Id to avoid having to fetch it.

While here, noticed that resolveCommits could panic if git log
returned an empty list, since we would Split the resulting empty string
and try to index into the missing space-separated fields, panicking.

I also noticed that we still asked the user to set up GITHUB_USER
and GITHUB_PAT, which hasn't been a strict requirement
since we added support for git credential helpers.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I529dfa5401625f1546971ed9520d0fe4f35d6a10