cmd/cueckoo: don't use fmt.Errorf to join errors
I was puzzled by a failure like:

    change "cue-labs%!F(MISSING)services~ma[...]" does not know about revision [...]

It turns out that we were stringifying errors, then re-joining them
with fmt.Errorf, meaning that the url-escaped slash "%2F"
was interpreted by fmt.Errorf as a missing parameter for the verb "F".

Luckily, we have an easy and shorter fix now via errors.Join,
which handles the nil case and joins via newlines like we were doing.

The error is now printed properly:

    change "cue-labs%2Fservices~ma[...]" does not know about revision [...]

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I685c6fecff2d923cd5cf13e20994958a1096f738