cmd/cueckoo: drop "thank you to" section
GitHub releases now show the users who contributed towards a release
at the bottom of the release UI, with their profile pictures,
driven by the usernames @-mentioned in the markdown release notes.

We already mention every contributor for every commit via the dropdown,
so we can drop the "thank you to" section and avoid duplication.

While here, make the commit list an expandable dropdown UI widget,
which I had been manually editing in previous releases.
It's easier for the tool to do this itself directly.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I72eec9e569c2f8a197b0f5ff139bf12e7c8041dc