internal/ci: switch to our namespace.so runner for testing
This is a custom runner profile I set up with all built-in forms of
caching enabled, including the git clone and Go's build and modules
caches. Note that the use of nscloud-checkout-action uses bind mounts,
so it is way faster to do the initial restore step, as well as the final
"post" step to store any results.

Change-Id: I7030e9eecfa0f2a5bc451634ecb629d18390f81e
Signed-off-by: Daniel Martí <mvdan@mvdan.cc>